+++ src/InstanceBuilder.ts
@@ -8,7 +8,7 @@ import isObject from './isObject';
 
 export type Constructor<T> = new (...args: any[]) => T;
 
-export default class InstanceBuilder<T, Struct extends object> extends Builder<T | null, Struct> {
+export default class InstanceBuilder<Struct extends object, T> extends Builder<Struct, T | null> {
   refClass: Constructor<T> | null = null;
 
   constructor(refClass: Constructor<T> | null = null) {
@@ -55,16 +55,16 @@ export default class InstanceBuilder<T, Struct extends object> extends Builder<T
   }
 }
 
-export function instance<T, S extends object>(
+export function instance<S extends object, T = Function>(
   refClass: Constructor<T> | null = null,
-): InstanceBuilder<T, S> {
-  return new InstanceBuilder(refClass);
+) /* infer */ {
+  return new InstanceBuilder<S, T>(refClass);
 }
 
-export function regex<S extends object>(): InstanceBuilder<RegExp, S> {
-  return instance(RegExp);
+export function regex<S extends object>() /* infer */ {
+  return instance<S, RegExp>(RegExp);
 }
 
-export function date<S extends object>(): InstanceBuilder<Date, S> {
-  return instance(Date);
+export function date<S extends object>() /* infer */ {
+  return instance<S, Date>(Date);
 }