+++ src/types.ts
@@ -4,20 +4,23 @@
  */
 
 import Builder from './Builder';
+import ArrayBuilder, { ArrayOf } from './ArrayBuilder';
+import ObjectBuilder, { ObjectOf } from './ObjectBuilder';
 
 // prettier-ignore
 export type Blueprint<Struct extends object> = {
   [K in keyof Struct]:
-    Struct[K] extends object ? Blueprint<Struct[K]> :
-    Struct[K] extends Builder<any, Struct> ? Struct[K] :
-    Builder<Struct[K], Struct>
+    // Struct[K] extends ArrayBuilder<infer S, infer T> ? ArrayBuilder<S, T> :
+    // Struct[K] extends ObjectBuilder<infer S, infer T> ? ObjectBuilder<S, T> :
+    Struct[K] extends Builder<infer S, infer T> ? Builder<S, T> :
+    Builder<Struct, Struct[K]>
 };
 
 // prettier-ignore
 export type InferStructure<Struct extends object> = {
   [K in keyof Struct]:
-    Struct[K] extends object ? InferStructure<Struct[K]> :
-    Struct[K] extends Builder<infer U, any> ? Infer<U> :
+    Struct[K] extends Builder<any, infer T> ? Infer<T> :
+    // Struct[K] extends object ? InferStructure<Struct[K]> :
     Struct[K]
 };
 
@@ -25,7 +28,7 @@ export type Infer<T> = T extends object ? InferStructure<T> : T;
 
 export type CheckerCallback = (path: string, value: any, ...args: any[]) => void;
 
-export type CustomCallback<Struct extends object> = (value: any, struct: Struct) => void;
+export type CustomCallback<Struct extends object = any> = (value: any, struct: Struct) => void;
 
 export interface OptimalOptions {
   file?: string;